+++ include/zix/sem.h
@@ -108,7 +108,8 @@ struct ZixSemImpl {
 static inline ZixStatus
 zix_sem_init(ZixSem* sem, unsigned val)
 {
-  return semaphore_create(mach_task_self(), &sem->sem, SYNC_POLICY_FIFO, val)
+  return semaphore_create(
+           mach_task_self(), &sem->sem, SYNC_POLICY_FIFO, (int)val)
            ? ZIX_STATUS_ERROR
            : ZIX_STATUS_SUCCESS;
 }
@@ -150,7 +151,7 @@ struct ZixSemImpl {
 static inline ZixStatus
 zix_sem_init(ZixSem* sem, unsigned initial)
 {
-  sem->sem = CreateSemaphore(NULL, initial, LONG_MAX, NULL);
+  sem->sem = CreateSemaphore(NULL, (LONG)initial, LONG_MAX, NULL);
   return (sem->sem) ? ZIX_STATUS_SUCCESS : ZIX_STATUS_ERROR;
 }
